Organize inputs for better tracking
I felt that the way I had the inputs before didn't allow my reflection
to have any bearing on my MIT for the next day, so I moved the MIT to
the end.

I also felt that I should be going from impulsive and emotional tracking
(mood and short journal entry) to thoughtful and deliberate tracking
(accomplishments, long journal) and then finally to actionable changes
for the next day (MIT).